Reorganized config structure
The config area is now much more logically structured. 

Also added support for specifying repo info via environment variables. This makes it easier to test this code against other repos or branches.
