+++ app.js
@@ -3,7 +3,7 @@
 const fs = require('fs');
 const process = require('process');
 const bodyParser = require('body-parser');
-const repo = require('./config/github-api').repo;
+const repo = require('./config/github/repo');
 const api = require('./lib/github-api');
 const featherIconHelper = require('./lib/helpers/feather-icon');
 const express = require('express');
