+++ config/github/index.js
@@ -1,0 +1,6 @@
+'use strict';
+
+const api = require('./api');
+const repo = require('./repo');
+
+module.exports = {api, repo};
