+++ config/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const github = require('./github');
+
+module.exports = {github};
