+++ lib/github-api/index.js
@@ -1,6 +1,6 @@
 'use strict';
 
 const GitHub = require('@octokit/rest');
-const apiConfig = require('../../config/github-api').apiConfig;
+const api = require('../../config/github/api');
 
-module.exports = GitHub(apiConfig);
+module.exports = GitHub(api);
